Squashed 'lispBM/lispBM/' changes from 747b679b..41f17e4a
41f17e4a changes to how errors are printed
7df9e408 small tweak repl

git-subtree-dir: lispBM/lispBM
git-subtree-split: 41f17e4a80816d2db079ecd9b3cb8ff372e5d6ef

+++ repl/lbm_sdl.c
@@ -378,7 +378,7 @@ static void blast_rgb565(uint8_t *dest, int dest_pitch, image_buffer_t *img) {
 
   uint32_t *w_dest = (uint32_t *)dest;
   for (int i = 0; i < num_pix; i ++) {
-    uint16_t pix = (((uint16_t)data[2 * i]) << 8) | ((uint16_t)data[2 * i + 1]);
+    uint16_t pix = (uint16_t)((((uint16_t)data[2 * i]) << 8) | ((uint16_t)data[2 * i + 1]));
 
     uint32_t r = (uint32_t)(pix >> 11);
     uint32_t g = (uint32_t)((pix >> 5) & 0x3F);
